requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
aiofiles
yt-dlp
requests
spotdl
spotipy
discord.py
python-dotenv
shazamio
pydub
mutagen
